♻️ Refactor: MessageCard component 분리
